implement renaming of libs before loading in for evaluation
