+++ Cargo.toml
@@ -33,6 +33,7 @@ libloading =	    { version = "0.5",	default-features = false }
 log =		    { version = "0.4",	default-features = false }
 racer =		    { version = "2.1",	default-features = false,   optional = true,	features = [ "metadata" ] }
 syn =		    { version = "1",	default-features = false,   optional = false,	features = [ "full", "printing", "parsing" ] }
+uuid =		    { version = "0.8",	default-features = false,   optional = false,	features = [ "v4" ] }
 
 [dev-dependencies]
 criterion = "0.3"
